requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pydantic
pydantic[email]
python-jose[cryptography]
python-multipart
cryptography
gunicorn


groq
motor
pymongo
